spec2chat/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
backports.tarfile==1.2.0
blis==0.7.11
build==1.2.2.post1
catalogue==2.0.10
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.16.0
confection==0.1.5
cymem==2.0.11
distro==1.9.0
dnspython==2.6.1
docutils==0.20.1
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
id==1.5.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
keyring==25.5.0
langcodes==3.4.1
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.5.0
murmurhash==1.0.12
nh3==0.2.21
nltk==3.9.1
numpy==1.24.4
openai==1.79.0
packaging==25.0
preshed==3.0.9
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pymongo==4.10.1
pyproject_hooks==1.2.0
python-dotenv==1.0.1
readme_renderer==43.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.0.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
-e git+https://github.com/mjesusrodriguez/spec2chat.git@6b01bcd11392ce371cf3d981dbc2d861dd1a25e0#egg=spec2chat
srsly==2.4.8
thinc==8.2.5
tomli==2.2.1
tqdm==4.67.1
twine==6.1.0
typer==0.9.4
typing_extensions==4.13.2
urllib3==2.2.3
wasabi==1.1.3
weasel==0.3.4
zipp==3.20.2
